Refactor: Move Application services interfaces to Common/Interfaces
Moved `INotificationService` and `IBatchJobProcessor` to `Valora.Application.Common.Interfaces`
to strictly adhere to the Clean Architecture constraint where the Application layer must define
interfaces but not implementations. Also updated referencing using directives to resolve
compilation issues.
