+++ backend/Valora.Api/Endpoints/NotificationEndpoints.cs
@@ -1,6 +1,6 @@
 using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
-using Valora.Application.Services;
+using Valora.Application.Common.Interfaces;
 using Valora.Domain.Enums;
 
 namespace Valora.Api.Endpoints;
