+++ backend/Valora.Application/Services/BatchJobs/MapGenerationJobProcessor.cs
@@ -1,6 +1,7 @@
 using Microsoft.Extensions.Logging;
 using Valora.Domain.Entities;
 using Valora.Domain.Enums;
+using Valora.Application.Common.Interfaces;
 
 namespace Valora.Application.Services.BatchJobs;
